:sparkles: Support server side stream gRPC transcoding via http event stream
